More type annotations

+++ AudioPlayer.py
@@ -1,16 +1,8 @@
 #!/usr/bin/env python
 
-"""extremely simple demonstration playing a soundfile
-and waiting for it to finish. you'll need the pygame.mixer
-module for this to work. Note how in this simple example we
-don't even bother loading all of the pygame package. Just
-pick the mixer for sound and time for the delay function.
-
-Optional command line argument:
-  the name of an audio file.
-  
-
-"""
+# Imports to support type annotations
+from __future__ import annotations
+from typing import Dict, Optional
 
 import os.path, sys, threading
 import pygame.mixer, pygame.time
@@ -19,47 +11,47 @@ time = pygame.time
 
 class AudioPlayer:
    class __AudioPlayer:
-      def __init__(self):
+      def __init__(self) -> None:
          #print ('in AudioPlayer::__AudioPlayer::__init__', flush=True)
 
          #choose a desired audio format
          mixer.init(11025) #raises exception on fail
          mixer.set_num_channels(32)
          
-         self.musicPath = None
-         self.soundPath = None
-         self.musicRelFilePath1 = None
-         self.musicRelFilePath2 = None
+         self.musicPath = './'
+         self.soundPath = './'
+         self.musicRelFilePath1: Optional[str] = None
+         self.musicRelFilePath2: Optional[str] = None
          self.running = True
-         self.sounds = {}
+         self.sounds: Dict[str, pygame.mixer.Sound] = {}
          self.musicThread = threading.Thread(target=self.__musicThread)
          self.musicThread.start()
          
-      def __del__(self):
+      def __del__(self) -> None:
          #print ('in AudioPlayer::__AudioPlayer::__del__', flush=True)
          self.terminate()
          self.musicThread.join()
          mixer.quit()
          
-      def setMusicPath(self, musicPath):
+      def setMusicPath(self, musicPath: str) -> None:
          #print ('in AudioPlayer::__AudioPlayer::setMusicPath', flush=True)
          self.musicPath = musicPath
          
-      def setSoundPath(self, soundPath):
+      def setSoundPath(self, soundPath: str) -> None:
          #print ('in AudioPlayer::__AudioPlayer::setSoundPath', flush=True)
          self.soundPath = soundPath
          
-      def playMusic(self, musicRelFilePath1, musicRelFilePath2 = None):
+      def playMusic(self, musicRelFilePath1: str, musicRelFilePath2: Optional[str] = None) -> None:
          #print ('in AudioPlayer::__AudioPlayer::playMusic', flush=True)
          self.musicRelFilePath1 = self.musicRelFilePath2 = musicRelFilePath1
          if musicRelFilePath2 is not None:
             self.musicRelFilePath2 = musicRelFilePath2
          
-      def __musicThread(self):
+      def __musicThread(self) -> None:
          #print ('in AudioPlayer::__AudioPlayer::__musicThread', flush=True)
          firstTime = True
-         currentMusicRelFilePath1 = None
-         currentMusicRelFilePath2 = None
+         currentMusicRelFilePath1: Optional[str] = None
+         currentMusicRelFilePath2: Optional[str] = None
          while self.running:
             # TODO: Switch to a more responsive approach which is not polling based
    
@@ -68,30 +60,8 @@ class AudioPlayer:
                currentMusicRelFilePath1 = self.musicRelFilePath1
                currentMusicRelFilePath2 = self.musicRelFilePath2
                firstTime = True
-            if self.musicRelFilePath1 != None and self.musicRelFilePath2 != None:
-
-               '''
-               #load the sound
-               if firstTime:
-                  sound = mixer.Sound( os.path.join(self.musicPath, self.musicRelFilePath1) )
-               else:
-                  sound = mixer.Sound( os.path.join(self.musicPath, self.musicRelFilePath2) )
-               firstTime = False
-
-               #start playing
-               #print ('Playing Sound...', flush=True)
-               channel = sound.play()
-
-               #poll until finished
-               while ( self.running and
-                       channel.get_busy() and
-                       currentMusicRelFilePath1 == self.musicRelFilePath1 and
-                       currentMusicRelFilePath2 == self.musicRelFilePath2 ): #still playing and not changed
-                  #print ('  ...still going...', flush=True)
-                  time.wait(100)
-               #print ('Stopping Sound...', flush=True)
-               channel.stop()'''
-
+               
+            if self.musicRelFilePath1 is not None and self.musicRelFilePath2 is not None:
                #load the music
                if firstTime:
                   mixer.music.load( os.path.join(self.musicPath, self.musicRelFilePath1) )
@@ -118,12 +88,12 @@ class AudioPlayer:
             time.wait(100)
          #print ('exitted AudioPlayer::__AudioPlayer::__musicThread', flush=True)
          
-      def playSound(self, soundRelFilePath):
+      def playSound(self, soundRelFilePath: str) -> None:
          #print ('in AudioPlayer::__AudioPlayer::playSound', flush=True)
          soundThread = threading.Thread( target=self.__soundThread, args=[soundRelFilePath] )
          soundThread.start()
          
-      def __soundThread(self, soundRelFilePath):
+      def __soundThread(self, soundRelFilePath: str) -> None:
          #print ('in AudioPlayer::__AudioPlayer::__soundThread', flush=True)
          # Load the sound if not previously loaded
          if not soundRelFilePath in self.sounds:
@@ -135,43 +105,46 @@ class AudioPlayer:
             time.wait(10)
          #print ('exitted AudioPlayer::__AudioPlayer::__soundThread', flush=True)
             
-      def stopMusic(self):
+      def stopMusic(self) -> None:
          #print ('in AudioPlayer::__AudioPlayer::stopMusic', flush=True)
          self.musicRelFilePath1 = self.musicRelFilePath2 = None
          
-      def terminate(self):
+      def terminate(self) -> None:
          #print ('in AudioPlayer::__AudioPlayer::terminate', flush=True)
          self.running = False
             
-   instance = None
-   def __init__(self):
+   instance: Optional[AudioPlayer.__AudioPlayer] = None
+   def __init__(self) -> None:
       if not AudioPlayer.instance:
          AudioPlayer.instance = AudioPlayer.__AudioPlayer()
          
-   def setMusicPath(self, musicPath):
-      self.instance.setMusicPath(musicPath)
+   def setMusicPath(self, musicPath: str) -> None:
+      if self.instance is not None:
+         self.instance.setMusicPath(musicPath)
       
-   def setSoundPath(self, soundPath):
-      self.instance.setSoundPath(soundPath)
+   def setSoundPath(self, soundPath: str) -> None:
+      if self.instance is not None:
+         self.instance.setSoundPath(soundPath)
       
-   def playMusic(self, musicFilePath1, musicFilePath2 = None):
-      self.instance.playMusic(musicFilePath1, musicFilePath2)
+   def playMusic(self, musicFilePath1: str, musicFilePath2: Optional[str] = None) -> None:
+      if self.instance is not None:
+         self.instance.playMusic(musicFilePath1, musicFilePath2)
       
-   def playSound(self, soundFilePath):
-      self.instance.playSound(soundFilePath)
+   def playSound(self, soundFilePath: str) -> None:
+      if self.instance is not None:
+         self.instance.playSound(soundFilePath)
       
-   def stopMusic(self):
-      self.instance.stopMusic()
+   def stopMusic(self) -> None:
+      if self.instance is not None:
+         self.instance.stopMusic()
       
-   def terminate(self):
-      self.instance.terminate()
-      self.instance.__del__()
-      self.instance = None
+   def terminate(self) -> None:
+      if self.instance is not None:
+         self.instance.terminate()
+         self.instance.__del__()
+         self.instance = None
 
-      
-
-
-def main():
+def main() -> None:
    
    # Initialize the music player
    audioPlayer = AudioPlayer()
